feat: create a menu option to list all movie names
Creates a new menu option 'list' that displays the title of all movies in the movie file, excluding the cast. The program uses a for loop to iterate through the movies list and takes only the name field from each dictionary to display to the user.
Moves code from display_menu.py to query.py.

issue #3