feat: create a menu option to search movie names
Creates a new menu option 'sn' which prompts the user to provide a word that they want to search for in movie names in the movie file. The program that takes that word and uses a for loop to iterate through the movie file to find movie names that contain the word given by the user, allowing for case insensitivity. If the condition is met, the program displays the movie name to the user.

issue #5