feat: create a menu option to search movie casts
Creates a new menu option 'sc' which prompts the user to provide a word that they want to search for in the movie cast in the movie file. The program takes that word and iterates through the cast list to find cast members names that contain the provided word. If the program finds that name, it displays the name of the movie and that cast members name to the user.

issue #7